revert dashboard components reorg, useQueryParams and TimeRangeProvider now inside dashboards package
