+++ ui/dashboards/src/components/index.ts
@@ -14,7 +14,6 @@
 export * from './Dashboard';
 export * from './GridLayout';
 export * from './Panel';
-export * from './PersesDashboard';
 export * from './TimeRangeControls';
 export * from './VariableAutocomplete';
 export * from './VariableList';
